Issue-7: Integrate with RDS
[Problem]
SQLite is not a production DB. It doesn't provide back-ups or the resiliency that is desired in a production
environment.

[Solution]
Add environment variables in the EBS environment, which are used to configure the Django DB if the environment variables
are defined.

[Test]
Deployed to EBS, ran through subscription flow, checked RDS backend that the data was there.