Issue-11: Create a view for a user to submit an alert
[Problem]
Users need a form to send alerts in.

[Solution]
Create a view that the user is able to specify the address information and integrate it with the Django app. Currently,
the action is just to print the "send" to STDOUT.

[Test]
Created test subscriber and sent notification via the alert form using local instance. Hopefully it works in prod.