Issue-19: Update main page to have welcome information
[Problem]
The main page had the default information from a RapidSMS installation. Additionally, people will need to be trained on
how to use the website, thus adding the basic information to the welcome page would be helpful.

[Solution]
Add basic information about the SMS Alert System and how to use the SMS Alert System website.

[Test]
* python manage.py runserver