Issue-20: Add additional languages
[Problem]
Mandarin, Korean, and Vietnamese are not supported languages in the system.

[Solution]
Add support for the languages by adding the required asset files and update the code to enable the languages.

[Test]
Tested using:
$ python manage.py runserver
and going through the Korean workflow. Added debug logging to print out the alert and follow-up notification to make
sure it was well formed.