Add WSGI alias for Let's Encrypt to grab the confirmation file
[Problem]
Let's encrypt requires the server to provide a file to it containing specific data. This file is served out of the host
directory:
http://rapidsms-alert-sys-env.h8iff8nwke.us-west-2.elasticbeanstalk.com/.well-known/acme-challenge/...

The directory is not automatically visible for Let's Encrypt.

[Solution]
Update the WSGI configuration to make the directory visible.

[Test]
Put a file in it and was successfully able to view it from the browser.